backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pypdf==3.14.0
pdf2image==1.16.3
pillow==10.0.1
pytesseract==0.3.13
python-multipart==0.0.6
aiofiles==23.1.0
python-dateutil==2.8.2

